Refactor PR assistant utilities into modules
